+++ sc/social/like/tests/test_canonicalurl_view.py
@@ -2,6 +2,7 @@
 """Test for the canonical URL updater form."""
 from DateTime import DateTime
 from plone import api
+from sc.social.like.interfaces import ISocialLikeSettings
 from sc.social.like.testing import HAS_DEXTERITY
 from sc.social.like.testing import INTEGRATION_TESTING
 from sc.social.like.tests.utils import enable_social_media_behavior
@@ -23,21 +24,30 @@ class CanonicalURLUpdaterTestCase(unittest.TestCase):
         self.portal = self.layer['portal']
         self.request = self.layer['request']
         enable_social_media_behavior()
+        self.setup_content()
+        api.portal.set_registry_record(
+            name='canonical_domain', value='https://example.org', interface=ISocialLikeSettings)
 
     def setup_content(self):
         with api.env.adopt_roles(['Manager']):
-            api.content.create(self.portal, type='News Item', id='foo')
-            api.content.create(self.portal, type='News Item', id='bar')
-            api.content.create(self.portal, type='News Item', id='baz')
+            obj = api.content.create(self.portal, type='News Item', id='foo')
+            api.content.transition(obj, 'publish')
+            obj = api.content.create(self.portal, type='News Item', id='bar')
+            api.content.transition(obj, 'publish')
+            obj = api.content.create(self.portal, type='News Item', id='baz')
+            api.content.transition(obj, 'publish')
 
         # simulate objects were create way long in the past
-        self.portal['foo'].creation_date = DateTime('2015/01/01')
-        self.portal['bar'].creation_date = DateTime('2016/01/01')
+        self.portal['foo'].effective_date = DateTime('2015/01/01')
         self.portal['foo'].reindexObject()
+        self.portal['bar'].effective_date = DateTime('2016/01/01')
         self.portal['bar'].reindexObject()
+        # XXX: publishing an object does not sets its effective date
+        #      https://github.com/plone/plone.api/issues/343
+        self.portal['baz'].effective_date = DateTime()
+        self.portal['baz'].reindexObject()
 
     def test_update_canonical_url(self):
-        self.setup_content()
         # canonical URL is None as we did not set up a canonical domain
         self.assertIsNone(self.portal['foo'].canonical_url)
         self.assertIsNone(self.portal['bar'].canonical_url)
@@ -47,16 +57,17 @@ class CanonicalURLUpdaterTestCase(unittest.TestCase):
         view = api.content.get_view(name, self.portal, self.request)
         # simulate data comming from form
         data = dict(
-            canonical_domain='https://example.org',
-            created_before=DateTime('2017/01/01').asdatetime(),
+            old_canonical_domain='http://example.org',
+            published_before=DateTime('2017/01/01').asdatetime(),
         )
         view.update_canonical_url(data)
         # objects created before the specified date will have their
         # canonical URL updated
         self.assertEqual(
-            self.portal['foo'].canonical_url, 'https://example.org/plone/foo')
+            self.portal['foo'].canonical_url, 'http://example.org/plone/foo')
         self.assertEqual(
-            self.portal['bar'].canonical_url, 'https://example.org/plone/bar')
+            self.portal['bar'].canonical_url, 'http://example.org/plone/bar')
         # objects created after the specified date will have their
         # canonical URL unchaged
-        self.assertIsNone(self.portal['baz'].canonical_url)
+        self.assertEqual(
+            self.portal['baz'].canonical_url, 'https://example.org/plone/baz')