+++ CHANGES.rst
@@ -3,8 +3,12 @@ Changelog
 
 There's a frood who really knows where his towel is.
 
-2.11 (unreleased)
-^^^^^^^^^^^^^^^^^
+2.10.1 (unreleased)
+^^^^^^^^^^^^^^^^^^^
+
+- Fix Canonical URL updater form;
+  a new upgrade step is provided to update the ``objects_provides`` catalog index (fixes `#115 <https://github.com/collective/sc.social.like/issues/115>`_).
+  [hvelarde]
 
 - Add ``canonical_domain`` field record to the registry when upgrading;
   this fixes an issue in the upgrade step to profile version 3045 (fixes `#114 <https://github.com/collective/sc.social.like/issues/114>`_).